requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
alembic==1.14.1
python-multipart==0.0.20
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1
authlib==1.3.2
itsdangerous==2.2.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.19.2
anthropic==0.40.0
